requirements.txt
fastapi 
uvicorn 
neo4j 
langchain-openai 
networkx 
numpy 
hnswlib 
python-dotenv